requirements.txt
aiogram>=3.0.0
pandas>=1.3.0
transformers>=4.12.0
torch>=1.9.0
python-dotenv>=0.19.0
aiosqlite>=0.17.0
Flask>=2.0.0
xlsxwriter>=3.0.0
requests>=2.25.0
aiohttp>=3.8.0
telethon
pyrogram
